Extract partition handling in custom handlers.
